metrics_service/requirements.txt
linux-metrics==0.1.4
python-daemon==2.2.3